+++ core/src/cli_commands.rs
@@ -2,9 +2,9 @@ use ergo_lib::ergotree_ir::chain::address::NetworkPrefix;
 
 pub mod bootstrap;
 pub mod extract_reward_tokens;
+pub mod prepare_update;
 pub mod print_reward_tokens;
 pub mod transfer_oracle_token;
-pub mod update;
 pub mod update_pool;
 pub mod vote_update_pool;
